requirements.txt
Pillow==10.0.1
blinker==1.8.2
boto3==1.28.1
MarkupSafe==2.1.5
botocore==1.31.85
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.15.4
Flask==2.3.3
fonttools==4.53.1
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.3
numpy==1.24.3
opencv-python==4.8.0.76
packaging==24.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
s3transfer==0.6.2
six==1.16.0
sympy==1.13.2
torch==2.0.1
typing-extensions==4.12.2
urllib3==2.0.7
Werkzeug==3.0.3
